Reorganises the network classes
