+++ src/implementation/Client/HTTPClient/HTTPClient.ts
@@ -23,8 +23,8 @@ import { Logger } from "../../../logger/Logger";
 import HTTPClientSidecar from "./sidecar";
 import { SDK_VERSION } from "../../../version";
 import * as SerializerUtil from "../../../utils/Serializer.util";
-import { HttpEndpoint } from "../../../utils/Client.util";
 import communicationProtocolEnum from "../../../enum/CommunicationProtocol.enum";
+import {HttpEndpoint} from "../../../network/HttpEndpoint";
 
 export default class HTTPClient implements IClient {
   readonly options: DaprClientOptions;
